Silence conversion error in Direction::to_index
Clang complains that the return statement cases from (signed) int to
(unsigned) size_t. As _val is {-1,0,1}, _val + 1 cannot be negative.

+++ flower/include/direction.hh
@@ -71,7 +71,7 @@ public:
   // convert to a zero-based index: {negative->0, zero->1, positive->2}
   constexpr size_t to_index () const
   {
-    return _val + 1;
+    return static_cast<size_t> (_val + 1);
   }
 
 private:
